+++ R/manage.R
@@ -16,7 +16,7 @@
 #' @export
 #' @examples
 #' \dontrun{
-#' # inventory of all downloaded data
+#' # inventory of all data downloaded to the default directory
 #' ebirdst_data_inventory()
 #'
 #' # inventory for a specific directory